% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppdx.R
\name{bin_bam}
\alias{bin_bam}
\title{Bin aligned sequences (from .bam) into genomic bins based on the .bed file.}
\usage{
bin_bam(bam_location, bed)
}
\arguments{
\item{bam_location}{A location to the BAM-file to bin.}

\item{bed}{A data frame in .bed format with columns: 'chr', 'start', 'end', 'focus'.}
}
\value{
A data frame in bed format with GC\%.
}
\description{
Bin aligned sequences (from .bam) into genomic bins based on the .bed file.
}
