% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppdx.R
\name{bin_counts}
\alias{bin_counts}
\title{Bin a BAM file for the bins provided by the .bed file.}
\usage{
bin_counts(bam_location, bed)
}
\arguments{
\item{bam_location}{A location to the BAM-file to bin.}

\item{bed}{A data frame in .bed format with columns: 'chr', 'start', 'end', 'focus'.}
}
\value{
A B data frame in bed format with GC\%.
}
\description{
This function calculates  z_score_PPDX_norm, z_score_PPDX and optionally
probabilities of being interest (recommended metric).
}
