% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calculate_bin_stat}
\alias{calculate_bin_stat}
\title{Calculate bin statistics. Can be only called right after normalization.}
\usage{
calculate_bin_stat(samples)
}
\arguments{
\item{samples}{A normalized data frame to be used in bin calculation.}
}
\value{
A data frame that has added columns PPDX, PPDX_norm-
}
\description{
Required columns "chr", "start", "reference", "gc_corrected".
}
\details{
\enumerate{
\item Calculates reference group each (by chromosome and start coordinate) bin
mean and sd.
\item Calculates regular Z-score for each sample (inc reference sample) and
if this samples bin is over ref mean bin.
\item Group by "sample", "chr", "focus", "reference" and calculate
}
\itemize{
\item PPDX (Z-score divided by sqrt())
\item PPDX_norm ((PPDX + 1) / (n_bins_over_mean + 2))
}
}
