% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_bam.R
\name{calculate_scores}
\alias{calculate_scores}
\title{Score BAM-file}
\usage{
calculate_scores(
  bam_location,
  reference_location,
  do_gc_correct = TRUE,
  use_pca = TRUE,
  nComp = 160,
  output_mahalanobis = TRUE,
  filter_on = FALSE,
  include_reference = FALSE,
  clean_env = TRUE
)
}
\arguments{
\item{bam_location}{Path to the input file}

\item{reference_location}{Path to the reference file}

\item{do_gc_correct}{Use bin-based GC-correct? PMID: 28500333 and PMID: 20454671}

\item{use_pca}{Use PCA based normalization? Has important effect for micro-deletions detection.}

\item{nComp}{How many components to use in PCA-based normalization. Should be lower than the number of genomic bins used.}

\item{output_mahalanobis}{Output also mahalanobis based probabilities of being different from the reference group?}

\item{include_reference}{Include reference in the output?}

\item{clean_env}{Optimize RAM usage by cleaning temp variables in this function?}

\item{bin_filter_on}{filter bins: filter(bin > mean(chr)), filter(bin_sd < mean(chr_sd)). Not recommended for detecting all the microdeletions. It can hide some CNVs, but can be helpful for detecting 45,X (if female fetus reference group is used.)}
}
\value{
A data frame with scores for the provided BAM.
}
\description{
This function calculates  z_score_PPDX_norm, z_score_PPDX and optionally
probabilities of being interest (recommended metric).
}
