% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calculate_summary}
\alias{calculate_summary}
\title{Calculate summary statistics (probabilities)}
\usage{
calculate_summary(samples)
}
\arguments{
\item{samples}{A binned statistics data frame.}
}
\value{
A data frame with probabilities
}
\description{
\enumerate{
\item Sums sample "PPDX", "PPDX_norm", "over_mean"
\item Calculates reference mean of "PPDX" and "PPDX_norm"
\item For each subregion (focus) calculates Mahalanobis distance from the
reference group (PPDX, PPDX_norm) and transforms it with chi-square(df = 2)
to -log10 probability.
}
}
