% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{chi_correct}
\alias{chi_correct}
\title{Chi-Squared Based Variation Reduction}
\usage{
chi_correct(samples, chi_cutoff = 3.5)
}
\description{
This function performs chi-squared based variation reduction to identify and correct overdispersed bins
}
