% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{gc_correct}
\alias{gc_correct}
\title{GC\% bin-correct (sample wise) (PMID: 28500333 and PMID: 20454671)}
\usage{
gc_correct(samples)
}
\arguments{
\item{samples}{A data frame to GC\% correct.}
}
\value{
A GC\% corrected data frame.
}
\description{
Expects columns "chr", "start", "end", "focus", "reads", "samples" and
creates column "gc_correct).
}
