% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindel.R
\name{infer_normality}
\alias{infer_normality}
\title{Infer sample high-risk probability.}
\usage{
infer_normality(
  bam_location,
  reference_location,
  use_pca = TRUE,
  cumulative_variance = 95,
  n_comp = NULL,
  bin_plot = FALSE,
  save_bins = FALSE
)
}
\arguments{
\item{bam_location}{Path to the input file.}

\item{reference_location}{Path to the reference file.}

\item{use_pca}{Use PCA based normalization? Increases sensitivity by reducing variation between samples.}

\item{cumulative_variance}{Choose number of PCA components based on the reference files such that number of PCA components explain x \% of cumulative variance.}

\item{n_comp}{Instead of automatically setting cumulative variance, set number of PCA components manually.  Cannot be higher than the number of samples or number of bins in the reference set.}

\item{bin_plot}{Create and save detailed bin plots?}

\item{save_bins}{Save bins?}
}
\description{
This function does the following:
\enumerate{
\item Apply bin-based GC\% correct.
\item Normalizing by read count.
\item Normalize by bin length.
\item Apply PCA-based normalization
\item Calculate reference group statistics per focus region.
\item Calculate sample Mahalanobis distance from the reference group.
\item Calculate -log10 chi-squared distribution probabilities.
}
}
\details{
Outputs results to different files.
}
\examples{

bam <- "sample.bam"
reference <- "reference.tsv"
infer_normality(bam, reference)
head("sample.bam.tsv")
}
