% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalize_reads}
\alias{normalize_reads}
\title{Normalize by read count and bin length. Expects GC\% correct to be done.}
\usage{
normalize_reads(samples)
}
\arguments{
\item{samples}{A data frame to normalize.}
}
\value{
A normalized data frame.
}
\description{
Normalize by read count and bin length. Expects GC\% correct to be done.
}
