% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{pca_correct}
\alias{pca_correct}
\title{PCA-normalize. Expects other normalization steps to be completed.
Based on:
https://stats.stackexchange.com/questions/229092/how-to-reverse-pca-and-reconstruct-original-variables-from-several-principal-com}
\usage{
pca_correct(samples)
}
\arguments{
\item{samples}{A data frame to PCA-normalize.}
}
\value{
A normalized data frame.
}
\description{
PCA-normalize. Expects other normalization steps to be completed.
Based on:
https://stats.stackexchange.com/questions/229092/how-to-reverse-pca-and-reconstruct-original-variables-from-several-principal-com
}
