% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niptdel.R
\name{plot_result}
\alias{plot_result}
\title{Plot inferred probabilities}
\usage{
plot_result(result)
}
\arguments{
\item{result}{inferred data frame}
}
\value{
A ggplot2 plot object
}
\description{
Plot inferred probabilities
}
