% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{save_result_plot}
\alias{save_result_plot}
\title{Create and save combined probability result figure.}
\usage{
save_result_plot(samples, sample_name)
}
\arguments{
\item{sample_name}{sample name of interest}

\item{data}{A probability data frame with columns "chr", "PPDX",
and "focus".}
}
\description{
Create and save combined probability result figure.
}
