% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niptdel.R
\name{find_gc}
\alias{find_gc}
\title{Find GC\% for GRCh38 based .bed data frame.}
\usage{
find_gc(bed)
}
\arguments{
\item{bed}{A data frame in .bed format with columns: 'chr', 'start', 'end', 'focus'.}
}
\value{
A data frame in bed format with GC\%.
}
\description{
Find GC\% for GRCh38 based .bed data frame.
}
