% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niptdel.R
\name{infer_normality}
\alias{infer_normality}
\title{Infer sample normality probability.}
\usage{
infer_normality(
  bam_location,
  reference_location,
  use_pca = TRUE,
  nComp = 80,
  bin_plot = TRUE,
  result_plot = TRUE,
  save_bins = FALSE
)
}
\arguments{
\item{bam_location}{Path to the input file}

\item{reference_location}{Path to the reference file}

\item{use_pca}{Use PCA based normalization?}

\item{nComp}{How many components to use in PCA-based normalization?}

\item{bin_plot}{Create and save detailed bin plots?}

\item{result_plot}{Create and save detailed result?}

\item{save_bins}{Save bins?}
}
\value{
A data frame with scores for the provided BAM.
}
\description{
This function does the following:
\enumerate{
\item Apply bin-based GC\% correct.
\item Normalizing by read count.
\item Normalize by bin length.
\item Apply PCA-based normalization
\item Calculate reference group statistics per focus region.
\item Calculate sample Mahalanobis distance from the reference group.
\item Calculate -log10 chi-squared distribution probabilities.
}
}
\examples{

bam <- "sample.bam"
reference <- "reference.tsv"
metrics <- infer_normality(bam, reference)
}
