% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppdx.R
\name{infer_normality}
\alias{infer_normality}
\title{Infer normality based on the binned read count.}
\usage{
infer_normality(
  bam_location,
  reference_location,
  do_gc_correct = TRUE,
  use_pca = TRUE,
  nComp = 160,
  output_mahalanobis = TRUE,
  bin_filter_on = FALSE,
  include_reference = FALSE,
  clean_env = TRUE
)
}
\arguments{
\item{bam_location}{Path to the input file}

\item{reference_location}{Path to the reference file}

\item{do_gc_correct}{Use bin-based GC-correct? PMID: 28500333 and PMID: 20454671}

\item{use_pca}{Use PCA based normalization? Has important effect for micro-deletions detection.}

\item{nComp}{How many components to use in PCA-based normalization. Should be lower than the number of genomic bins used.}

\item{bin_filter_on}{filter bins: filter(bin > mean(chr)), filter(bin_sd < mean(chr_sd)). Not recommended for detecting all the microdeletions. It can hide some CNVs, but can be helpful for detecting 45,X (if female fetus reference group is used).}

\item{include_reference}{Include reference in the output?}

\item{clean_env}{Optimize RAM usage by cleaning temp variables in this function?}
}
\value{
A data frame with scores for the provided BAM.
}
\description{
Infer normality based on the binned read count.
}
